+++ OzzTradeDiary.NET/OzzTradeDiary.SQLite/DbScripts/EntryOrder.sql
@@ -8,8 +8,9 @@ CREATE TABLE IF NOT EXISTS EntryOrders(
 	FilledQuantity TEXT, 
 	OrderValue INTEGER, 
 	FilledValue INTEGER, 
-	FilledTime INTEGER, 
 	Notes TEXT, 
+	CancellationTime TEXT, 
+	FilledTime TEXT, 
 	UpdatedAt TEXT Not Null 
 );
 Create Index If Not Exists idx_EntryOrders_FilledTime on EntryOrders(FilledTime);